refactor: move Peer to its own file
