+++ hyper-lib/src/lib.rs
@@ -7,6 +7,7 @@ use statistics::NetworkStatistics;
 pub mod graph;
 pub mod network;
 pub mod node;
+pub mod peer;
 pub mod simulator;
 pub mod statistics;
 pub mod txreconciliation;
